+++ packages/keyring-controller/tests/mocks/mockMutableKeyring.ts
@@ -1,0 +1,19 @@
+import { MockKeyring } from './mockKeyring';
+
+export class MockMutableKeyring extends MockKeyring {
+  static type = 'Mock Mutable Keyring';
+
+  public type = 'Mock Mutable Keyring';
+
+  #updated: boolean = false;
+
+  update() {
+    this.#updated = true;
+  }
+
+  async serialize() {
+    return Promise.resolve({
+      updated: this.#updated,
+    });
+  }
+}